+++ test/server/src/index.test.ts
@@ -6,7 +6,7 @@ import { Location, Range } from 'vscode-languageserver/node';
 import { 
 	Index,
 	FlowControlEvent,
-	IdentifierMultiIndex, 
+	IdentifierMultiIndex,
 } from '../../../server/src/index';
 
 const documentUri = "file:///faker.txt";
@@ -33,9 +33,10 @@ describe("Project Index", () => {
 			expect(references[1].range.end).to.eql({line: 2, character: 5});
 		});
 
-		it("should return choicescript_stats in the scene list if present in the index", () => {
+		it("should only return choicescript_stats once in the scene list if present in the index", () => {
 			const index = new Index();
 			index.setHasChoicescriptStats(true);
+			index.setSceneList(["choicescript_stats"]);
 
 			const scenes = index.getSceneList();
 
@@ -104,6 +105,97 @@ describe("Project Index", () => {
 			expect(references[1].range.end).to.eql({line: 7, character: 7});
 		});
 
+		it("should find scene references across all goto- and gosub-scene", () => {
+			const localEvents: FlowControlEvent[] = [
+				{
+					command: "goto_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "away_label",
+					labelLocation: Location.create(documentUri, Range.create(2, 0, 2, 5)),
+					scene: "scene1"
+				},
+				{
+					command: "gosub_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "other_label",
+					labelLocation: Location.create(documentUri, Range.create(2, 0, 2, 5)),
+					scene: "scene2"
+				}
+			];
+			const otherSceneEvents: FlowControlEvent[] = [
+				{
+					command: "gosub_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "local_label",
+					labelLocation: Location.create(otherSceneUri, Range.create(7, 0, 7, 7)),
+					scene: "scene1",
+					sceneLocation: Location.create(otherSceneUri, Range.create(8, 0, 8, 8))
+				}
+			];
+			const index = new Index();
+			index.setFlowControlEvents(documentUri, localEvents);
+			index.setFlowControlEvents(otherSceneUri, otherSceneEvents);
+
+			const scenes = index.getAllReferencedScenes();
+
+			expect(scenes).to.eql(["scene1", "scene2"]);
+		});
+
+		it("should not find scene references that are themselves variable references", () => {
+			const localEvents: FlowControlEvent[] = [
+				{
+					command: "goto_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "away_label",
+					labelLocation: Location.create(documentUri, Range.create(2, 0, 2, 5)),
+					scene: "scene1"
+				},
+				{
+					command: "gosub_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "other_label",
+					labelLocation: Location.create(documentUri, Range.create(2, 0, 2, 5)),
+					scene: "scene2"
+				}
+			];
+			const otherSceneEvents: FlowControlEvent[] = [
+				{
+					command: "gosub_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "local_label",
+					labelLocation: Location.create(otherSceneUri, Range.create(7, 0, 7, 7)),
+					scene: "{var1}",
+					sceneLocation: Location.create(otherSceneUri, Range.create(8, 0, 8, 8))
+				}
+			];
+			const index = new Index();
+			index.setFlowControlEvents(documentUri, localEvents);
+			index.setFlowControlEvents(otherSceneUri, otherSceneEvents);
+
+			const scenes = index.getAllReferencedScenes();
+
+			expect(scenes).to.eql(["scene1", "scene2"]);
+		});
+
+		it("should include explicitly listed scenes in the scene references", () => {
+			const localEvents: FlowControlEvent[] = [
+				{
+					command: "goto_scene",
+					commandLocation: Substitute.for<Location>(),
+					label: "away_label",
+					labelLocation: Location.create(documentUri, Range.create(2, 0, 2, 5)),
+					scene: "scene1"
+				}
+			];
+			const index = new Index();
+			index.setFlowControlEvents(documentUri, localEvents);
+			index.setSceneList(["startup"]);
+
+			const scenes = index.getAllReferencedScenes();
+
+			expect(scenes).to.eql(["startup", "scene1"]);
+		});
+
 		it("should find achievement references in a scene", () => {
 			const referenceLocation = Location.create(documentUri, Range.create(1, 1, 1, 7));
 			const achievementReferences = new Map([["achieve", [referenceLocation]]]);
@@ -135,6 +227,101 @@ describe("Project Index", () => {
 			expect(references[0].range.end).to.eql({line: 1, character: 7});
 			expect(references[1].range.start).to.eql({line: 9, character: 1});
 			expect(references[1].range.end).to.eql({line: 9, character: 7});
+		});
+
+		describe("Indexed Scenes", () => {
+			// Since indexed scenes come from so many sources, group them
+			it("should include word count URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setWordCount(documentUri, 0);
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include local variable URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setLocalVariables(documentUri, new Map());
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include subroutine local varable URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setSubroutineLocalVariables(documentUri, new Map());
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include variable reference URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setVariableReferences(documentUri, new Map());
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include label URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setLabels(documentUri, new Map());
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include achievement reference URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setAchievementReferences(documentUri, new Map());
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include document scope URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setDocumentScopes(documentUri, { achievementVarScopes: [], choiceScopes: [], paramScopes: []});
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include flow control event URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setFlowControlEvents(documentUri, []);
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should include parse error URIs in indexed scenes", () => {
+				const index = new Index();
+				index.setParseErrors(documentUri, []);
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker"]);
+			});
+
+			it("should not return scenes more than once in the indexed scenes", () => {
+				const index = new Index();
+				index.setWordCount(documentUri, 0);
+				index.setParseErrors(documentUri, []);
+				index.setParseErrors(otherSceneUri, []);
+
+				const scenes = index.getIndexedScenes();
+
+				expect(scenes).to.eql(["faker", "other-scene"]);
+			});
 		});
 	});
 });