+++ test/server/src/indexer.test.ts
@@ -27,12 +27,19 @@ function createDocument(text: string,
 	return fakeDocument;
 }
 
+function createIndex(): SubstituteOf<ProjectIndex> {
+	let fakeIndex = Substitute.for<ProjectIndex>();
+	fakeIndex.getIndexedScenes().returns([]);
+	fakeIndex.getAllReferencedScenes().returns([]);
+	return fakeIndex;
+}
+
 describe("Indexer", () => {
 	describe("Symbol Creation Indexing", () => {
 		it("should index locations of created global variables", () => {
 			let fakeDocument = createDocument("*create variable 3");
 			let received: Array<IdentifierIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setGlobalVariables(Arg.all()).mimicks((uri: string, index: IdentifierIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -45,7 +52,7 @@ describe("Indexer", () => {
 		it("should index locations of created local variables", () => {
 			let fakeDocument = createDocument("*temp variable 3");
 			let received: Array<IdentifierMultiIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLocalVariables(Arg.all()).mimicks((uri: string, index: IdentifierMultiIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -59,7 +66,7 @@ describe("Indexer", () => {
 		it("should index all created locations local variables", () => {
 			let fakeDocument = createDocument("*temp variable 3\n*temp variable 1");
 			let received: Array<IdentifierMultiIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLocalVariables(Arg.all()).mimicks((uri: string, index: IdentifierMultiIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -75,7 +82,7 @@ describe("Indexer", () => {
 		it("should index effective locations of local variables created in a subroutine", () => {
 			let fakeDocument = createDocument("*gosub subroutine\n*finish\n*label subroutine\n*temp variable 3\n*return\n");
 			let received: Array<IdentifierIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setSubroutineLocalVariables(Arg.all()).mimicks((uri: string, index: IdentifierIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -88,7 +95,7 @@ describe("Indexer", () => {
 		it("should not index effective locations of local variables created in a subroutine at a gosub that's after the subroutine", () => {
 			let fakeDocument = createDocument("*label s\n*params v1\n*set {v1} 2\n*return\n\n*gosub s 1");
 			let received: Array<IdentifierIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setSubroutineLocalVariables(Arg.all()).mimicks((uri: string, index: IdentifierIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -100,7 +107,7 @@ describe("Indexer", () => {
 		it("should not re-index effective locations of local variables redefined in a subroutine called by a previous subroutine", () => {
 			let fakeDocument = createDocument("*label sub1\n*temp s1 1\n*set s1 2\n*gosub sub2\n*return\n*label sub2\n*temp s1 3\n*return");
 			let received: Array<IdentifierIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setSubroutineLocalVariables(Arg.all()).mimicks((uri: string, index: IdentifierIndex) => { received.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -114,7 +121,7 @@ describe("Indexer", () => {
 		it("should index bare variables in the command", () => {
 			let fakeDocument = createDocument("*set variable 3");
 			let receivedReferences: Array<IdentifierMultiIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setVariableReferences(Arg.all()).mimicks((uri: string, index: IdentifierMultiIndex) => { receivedReferences.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -128,7 +135,7 @@ describe("Indexer", () => {
 		it("should index variables in the test", () => {
 			let fakeDocument = createDocument("@{variable this | that}");
 			let receivedReferences: Array<IdentifierMultiIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setVariableReferences(Arg.all()).mimicks((uri: string, index: IdentifierMultiIndex) => { receivedReferences.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -142,7 +149,7 @@ describe("Indexer", () => {
 		it("should index local variables directly after a reference command", () => {
 			let fakeDocument = createDocument("*if variable > 1");
 			let receivedReferences: Array<IdentifierMultiIndex> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setVariableReferences(Arg.all()).mimicks((uri: string, index: IdentifierMultiIndex) => { receivedReferences.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -156,7 +163,7 @@ describe("Indexer", () => {
 		it("should capture the flow control command", () => {
 			let fakeDocument = createDocument("*gosub label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -169,7 +176,7 @@ describe("Indexer", () => {
 		it("should capture the label", () => {
 			let fakeDocument = createDocument("*gosub label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -180,7 +187,7 @@ describe("Indexer", () => {
 		it("should capture the label location", () => {
 			let fakeDocument = createDocument("*gosub label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -192,7 +199,7 @@ describe("Indexer", () => {
 		it("should skip the scene when not present", () => {
 			let fakeDocument = createDocument("*gosub label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -204,7 +211,7 @@ describe("Indexer", () => {
 		it("should capture the scene", () => {
 			let fakeDocument = createDocument("*gosub_scene scene-name");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -215,7 +222,7 @@ describe("Indexer", () => {
 		it("should capture the scene location", () => {
 			let fakeDocument = createDocument("*gosub_scene scene-name");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -227,7 +234,7 @@ describe("Indexer", () => {
 		it("should skip the label when not present", () => {
 			let fakeDocument = createDocument("*gosub_scene scene-name");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -239,7 +246,7 @@ describe("Indexer", () => {
 		it("should capture both scene and label if present", () => {
 			let fakeDocument = createDocument("*gosub_scene scene-name other_label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -251,7 +258,7 @@ describe("Indexer", () => {
 		it("should capture both scene and label locations if present", () => {
 			let fakeDocument = createDocument("*gosub_scene scene-name other_label");
 			let receivedReferences: Array<FlowControlEvent[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setFlowControlEvents(Arg.all()).mimicks((uri: string, events: FlowControlEvent[]) => { receivedReferences.push(events); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -267,7 +274,7 @@ describe("Indexer", () => {
 		it("should index scenes in startup files", () => {
 			let fakeDocument = createDocument("*scene_list\n\tstartup\n\tscene-1\n\tscene-2\n");
 			let receivedScenes: Array<Array<string>> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setSceneList(Arg.any()).mimicks((scenes: string[]) => { receivedScenes.push(scenes); });
 			fakeIndex.hasChoicescriptStats().returns(false);
 
@@ -276,34 +283,10 @@ describe("Indexer", () => {
 			expect(receivedScenes).to.eql([['startup', 'scene-1', 'scene-2']]);
 		});
 
-		it("should include the startup scene even if it's not listed in the startup file", () => {
-			let fakeDocument = createDocument("*scene_list\n\tscene-1\n\tscene-2\n");
-			let receivedScenes: Array<Array<string>> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
-			fakeIndex.setSceneList(Arg.any()).mimicks((scenes: string[]) => { receivedScenes.push(scenes); });
-			fakeIndex.hasChoicescriptStats().returns(false);
-
-			updateProjectIndex(fakeDocument, true, false, fakeIndex);
-
-			expect(receivedScenes).to.eql([['scene-1', 'scene-2', 'startup']]);
-		});
-
-		it("should include the choicescript_stats scene even if it's not listed in the startup file", () => {
-			let fakeDocument = createDocument("*scene_list\n\tscene-1\n\tscene-2\n");
-			let receivedScenes: Array<Array<string>> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
-			fakeIndex.setSceneList(Arg.any()).mimicks((scenes: string[]) => { receivedScenes.push(scenes); });
-			fakeIndex.hasChoicescriptStats().returns(true);
-
-			updateProjectIndex(fakeDocument, true, false, fakeIndex);
-
-			expect(receivedScenes).to.eql([['scene-1', 'scene-2', 'startup', 'choicescript_stats']]);
-		});
-
 		it("should set the hasChoicescriptStats flag if it indexes a choicescript_stats scene", () => {
 			let fakeDocument = createDocument("*comment\n");
 			let receivedScenes: Array<Array<string>> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setSceneList(Arg.any()).mimicks((scenes: string[]) => { receivedScenes.push(scenes); });
 			fakeIndex.getSceneList().returns([]);
 
@@ -312,21 +295,9 @@ describe("Indexer", () => {
 			fakeIndex.received(1).setHasChoicescriptStats(true);
 		});
 
-		it("should add the choicescript_stats scene to the scene list even if it's not listed in the startup file", () => {
-			let fakeDocument = createDocument("*comment\n");
-			let receivedScenes: Array<Array<string>> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
-			fakeIndex.setSceneList(Arg.any()).mimicks((scenes: string[]) => { receivedScenes.push(scenes); });
-			fakeIndex.getSceneList().returns(['old_scene']);
-
-			updateProjectIndex(fakeDocument, false, true, fakeIndex);
-
-			expect(receivedScenes).to.eql([['old_scene', 'choicescript_stats']]);
-		});
-
 		it("should not add the choicescript_stats scene to the scene list if it's already listed in the startup file", () => {
 			let fakeDocument = createDocument("*comment\n");
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.getSceneList().returns(['choicescript_stats', 'old_scene']);
 
 			updateProjectIndex(fakeDocument, false, true, fakeIndex);
@@ -339,7 +310,7 @@ describe("Indexer", () => {
 		it("should set has-choicescript-stats for stats files", () => {
 			let fakeDocument = createDocument("");
 			let receivedCalls: Array<boolean> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setHasChoicescriptStats(Arg.any()).mimicks((canHas: boolean) => { receivedCalls.push(canHas); })
 			fakeIndex.getSceneList().returns([]);
 
@@ -351,7 +322,7 @@ describe("Indexer", () => {
 		it("should not set has-choicescript-stats for non-stats files", () => {
 			let fakeDocument = createDocument("");
 			let receivedCalls: Array<boolean> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setHasChoicescriptStats(Arg.any()).mimicks((canHas: boolean) => { receivedCalls.push(canHas); })
 
 			updateProjectIndex(fakeDocument, false, false, fakeIndex);
@@ -364,7 +335,7 @@ describe("Indexer", () => {
 		it("should index an achievement", () => {
 			let fakeDocument = createDocument("*achievement code_name");
 			let receivedAchievements: IdentifierIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setAchievements(Arg.any()).mimicks((index: IdentifierIndex) => { receivedAchievements.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -376,7 +347,7 @@ describe("Indexer", () => {
 		it("should index a reference to an achievement", () => {
 			let fakeDocument = createDocument("*achieve code_name");
 			let receivedAchievementReferences: IdentifierMultiIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setAchievementReferences(Arg.any(), Arg.any()).mimicks(
 				(uri, index) => {
 					receivedAchievementReferences.push(index);
@@ -394,7 +365,7 @@ describe("Indexer", () => {
 		it("should index a label", () => {
 			let fakeDocument = createDocument("*label label_name");
 			let receivedLabels: LabelIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLabels(Arg.any(), Arg.any()).mimicks((uri: string, index: LabelIndex) => { receivedLabels.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -406,7 +377,7 @@ describe("Indexer", () => {
 		it("should leave a label's scope blank on no return", () => {
 			let fakeDocument = createDocument("*label label_name");
 			let receivedLabels: LabelIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLabels(Arg.any(), Arg.any()).mimicks((uri: string, index: LabelIndex) => { receivedLabels.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -418,7 +389,7 @@ describe("Indexer", () => {
 		it("should add a scope to a label on a *return", () => {
 			let fakeDocument = createDocument("*label label_name\nLine 1\n*return\nLine 3");
 			let receivedLabels: LabelIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLabels(Arg.any(), Arg.any()).mimicks((uri: string, index: LabelIndex) => { receivedLabels.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -431,7 +402,7 @@ describe("Indexer", () => {
 		it("should add a scope to the most recent label on a *return", () => {
 			let fakeDocument = createDocument("*label label_one\nLine 1\n*label label_two\nLine 3\n*return\nLine 5");
 			let receivedLabels: LabelIndex[] = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setLabels(Arg.any(), Arg.any()).mimicks((uri: string, index: LabelIndex) => { receivedLabels.push(index); });
 
 			updateProjectIndex(fakeDocument, true, false, fakeIndex);
@@ -447,7 +418,7 @@ describe("Indexer", () => {
 		it("should capture *check_achievements scopes", () => {
 			let fakeDocument = createDocument("Line 0\n*check_achievements\nLine 2");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -461,7 +432,7 @@ describe("Indexer", () => {
 		it("should have *check_achievements scope run from definition to document end", () => {
 			let fakeDocument = createDocument("Line 0\n*check_achievements\nLine 2\nLast line");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -476,7 +447,7 @@ describe("Indexer", () => {
 		it("should capture *params scopes", () => {
 			let fakeDocument = createDocument("Line 0\n*params\nLine 2");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -490,7 +461,7 @@ describe("Indexer", () => {
 		it("should have *params scope run from definition to document end", () => {
 			let fakeDocument = createDocument("Line 0\n*params\nLine 2\nLast line");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -507,7 +478,7 @@ describe("Indexer", () => {
 		it("should capture the scope of a *label followed by a *return", () => {
 			let fakeDocument = createDocument("Line 0\n*label\nLine 2\n*return\nLine 4");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -521,7 +492,7 @@ describe("Indexer", () => {
 		it("should have *label scope run from *label to *return", () => {
 			let fakeDocument = createDocument("Line 0\n*label\nLine 2\n*return\nLine 4");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -538,7 +509,7 @@ describe("Indexer", () => {
 		it("should capture a *choice block", () => {
 			let fakeDocument = createDocument("Line 0\n*choice\n    #One\n        Text\n    #Two\nEnd");
 			let received: Array<DocumentScopes> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setDocumentScopes(Arg.all()).mimicks(
 				(uri: string, scope: DocumentScopes) => { received.push(scope); }
 			);
@@ -553,7 +524,7 @@ describe("Indexer", () => {
 		it("should flag attempts to re-create already-created global variables", () => {
 			let fakeDocument = createDocument("*create variable 3\n*create variable 9");
 			let received: Array<Diagnostic[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setParseErrors(Arg.all()).mimicks(
 				(uri: string, errors: Diagnostic[]) => { received.push(errors); }
 			);
@@ -570,7 +541,7 @@ describe("Indexer", () => {
 		it("should flag attempts to re-create already-created labels", () => {
 			let fakeDocument = createDocument("*label previous_label\n*label previous_label");
 			let received: Array<Diagnostic[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setParseErrors(Arg.all()).mimicks(
 				(uri: string, errors: Diagnostic[]) => { received.push(errors); }
 			);
@@ -587,7 +558,7 @@ describe("Indexer", () => {
 		it("should flag a *return with no label", () => {
 			let fakeDocument = createDocument("*return");
 			let received: Array<Diagnostic[]> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setParseErrors(Arg.all()).mimicks(
 				(uri: string, errors: Diagnostic[]) => { received.push(errors); }
 			);
@@ -606,7 +577,7 @@ describe("Indexer", () => {
 		it("should count words in a document", () => {
 			let fakeDocument = createDocument("*create variable 3\nIt's three words\n*comment don't count me");
 			let received: Array<number> = [];
-			let fakeIndex = Substitute.for<ProjectIndex>();
+			let fakeIndex = createIndex();
 			fakeIndex.setWordCount(Arg.all()).mimicks(
 				(uri: string, count: number) => { received.push(count); }
 			);
@@ -615,6 +586,20 @@ describe("Indexer", () => {
 
 			expect(received.length).to.equal(1);
 			expect(received[0]).to.equal(3);
+		});
+	});
+
+	describe("Non-Indexed Scenes", () => {
+		it("should return scenes that haven't yet been indexed", () => {
+			let fakeDocument = createDocument("");
+			let received: Array<number> = [];
+			let fakeIndex = Substitute.for<ProjectIndex>();
+			fakeIndex.getIndexedScenes().returns(["startup"]);
+			fakeIndex.getAllReferencedScenes().returns(["startup", "new-scene"]);
+
+			const newScenes = updateProjectIndex(fakeDocument, true, false, fakeIndex);
+
+			expect(newScenes).to.eql(["new-scene"]);
 		});
 	});
 });